[DBSCAN] Re-organized some methods and added the contents of tasks.py to base.py
